refactor: extract UI overlays from CoreUI into separate components
Move full-screen overlay components into dedicated files for better
maintainability and reusability:

- DisconnectedOverlay: Auto-reconnect with countdown and manual retry
- KickedOverlay: Display kick reason messages (duplicate_user, player_limit)
- DeathScreen: Death UI with respawn button, kill info, and item timer

This reduces CoreUI.tsx by ~300 lines and creates a modular overlay
system that can be easily extended for future game states.
