+++ packages/client/src/game/CoreUI.tsx
@@ -1,4 +1,3 @@
-import { RefreshCwIcon } from "lucide-react";
 import React, { useEffect, useRef, useState } from "react";
 import { useThemeStore } from "@/ui";
 
@@ -23,6 +22,7 @@ import { LevelUpNotification } from "./hud/level-up";
 import { EscapeMenu } from "./hud/EscapeMenu";
 import { ConnectionIndicator } from "./hud/ConnectionIndicator";
 import { NotificationContainer } from "@/ui/components";
+import { Disconnected, KickedOverlay, DeathScreen } from "./hud/overlays";
 import {
   COLORS,
   spacing,
@@ -417,323 +417,6 @@ export function CoreUI({ world }: { world: ClientWorld }) {
         {deathScreen && <DeathScreen data={deathScreen} world={world} />}
       </main>
     </ChatProvider>
-  );
-}
-
-function Disconnected() {
-  const theme = useThemeStore((s) => s.theme);
-  const [countdown, setCountdown] = useState(5);
-  const [isAutoReconnecting, setIsAutoReconnecting] = useState(true);
-
-  // Auto-reconnect countdown
-  useEffect(() => {
-    if (!isAutoReconnecting || countdown <= 0) return;
-
-    const timer = setInterval(() => {
-      setCountdown((prev) => {
-        if (prev <= 1) {
-          clearInterval(timer);
-          window.location.reload();
-          return 0;
-        }
-        return prev - 1;
-      });
-    }, 1000);
-
-    return () => clearInterval(timer);
-  }, [isAutoReconnecting, countdown]);
-
-  const handleManualReconnect = () => {
-    window.location.reload();
-  };
-
-  const handleCancelAutoReconnect = () => {
-    setIsAutoReconnecting(false);
-  };
-
-  return (
-    <>
-      <div className="fixed top-0 left-0 w-full h-full backdrop-grayscale pointer-events-none z-[9999] opacity-0 animate-[fadeIn_3s_ease-in-out_forwards]" />
-      <style>{`
-        @keyframes fadeIn {
-          from { opacity: 0; }
-          to { opacity: 1; }
-        }
-        @keyframes pulse {
-          0%, 100% { opacity: 1; }
-          50% { opacity: 0.5; }
-        }
-      `}</style>
-      <div
-        className="pointer-events-auto absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 backdrop-blur-md rounded-2xl p-4 flex flex-col items-center gap-3"
-        style={{
-          backgroundColor: theme.colors.background.secondary,
-          border: `1px solid ${theme.colors.border.default}`,
-          color: theme.colors.text.primary,
-          minWidth: "240px",
-        }}
-      >
-        <div className="flex items-center gap-2">
-          <div
-            className="w-2 h-2 rounded-full"
-            style={{
-              backgroundColor: isAutoReconnecting ? "#f59e0b" : "#ef4444",
-              animation: isAutoReconnecting
-                ? "pulse 1.5s ease-in-out infinite"
-                : "none",
-            }}
-          />
-          <span className="font-medium">Connection Lost</span>
-        </div>
-
-        {isAutoReconnecting ? (
-          <>
-            <div className="text-sm opacity-70">
-              Reconnecting in {countdown}s...
-            </div>
-            <div className="flex gap-2">
-              <button
-                className="px-3 py-1.5 rounded-lg flex items-center gap-2 cursor-pointer transition-colors"
-                style={{
-                  backgroundColor:
-                    theme.colors.accent?.primary || theme.colors.border.default,
-                  color: theme.colors.text.primary,
-                }}
-                onClick={handleManualReconnect}
-              >
-                <RefreshCwIcon size={14} />
-                <span>Reconnect Now</span>
-              </button>
-              <button
-                className="px-3 py-1.5 rounded-lg cursor-pointer transition-opacity hover:opacity-80"
-                style={{
-                  backgroundColor: "transparent",
-                  border: `1px solid ${theme.colors.border.default}`,
-                  color: theme.colors.text.secondary,
-                }}
-                onClick={handleCancelAutoReconnect}
-              >
-                Cancel
-              </button>
-            </div>
-          </>
-        ) : (
-          <>
-            <div className="text-sm opacity-70">Auto-reconnect cancelled</div>
-            <button
-              className="px-4 py-2 rounded-lg flex items-center gap-2 cursor-pointer transition-colors"
-              style={{
-                backgroundColor:
-                  theme.colors.accent?.primary || theme.colors.border.default,
-                color: theme.colors.text.primary,
-              }}
-              onClick={handleManualReconnect}
-            >
-              <RefreshCwIcon size={16} />
-              <span>Reconnect</span>
-            </button>
-          </>
-        )}
-      </div>
-    </>
-  );
-}
-
-const kickMessages: Record<string, string> = {
-  duplicate_user: "Player already active on another device or window.",
-  player_limit: "Player limit reached.",
-  unknown: "You were kicked.",
-};
-function KickedOverlay({ code }: { code: string }) {
-  const theme = useThemeStore((s) => s.theme);
-  return (
-    <div
-      className="absolute inset-0 flex items-center justify-center pointer-events-auto"
-      style={{ backgroundColor: theme.colors.background.primary }}
-    >
-      <div className="text-lg" style={{ color: theme.colors.text.primary }}>
-        {kickMessages[code] || kickMessages.unknown}
-      </div>
-    </div>
-  );
-}
-
-function DeathScreen({
-  data,
-  world,
-}: {
-  data: { message: string; killedBy: string; respawnTime: number };
-  world: ClientWorld;
-}) {
-  const theme = useThemeStore((s) => s.theme);
-  // Track respawn state to prevent button spam
-  const [isRespawning, setIsRespawning] = useState(false);
-  // Track if respawn request timed out
-  const [respawnTimedOut, setRespawnTimedOut] = useState(false);
-  // Death countdown timer - seconds until items despawn
-  const [countdown, setCountdown] = useState<number>(
-    Math.max(0, Math.floor((data.respawnTime - Date.now()) / 1000)),
-  );
-
-  // Timeout handler - re-enable button if server doesn't respond
-  const RESPAWN_TIMEOUT_MS = 10000; // 10 seconds
-
-  useEffect(() => {
-    if (!isRespawning) return;
-
-    const timeoutId = setTimeout(() => {
-      console.warn("[DeathScreen] Respawn request timed out after 10 seconds");
-      setIsRespawning(false);
-      setRespawnTimedOut(true);
-    }, RESPAWN_TIMEOUT_MS);
-
-    return () => clearTimeout(timeoutId);
-  }, [isRespawning]);
-
-  // Update countdown every second
-  useEffect(() => {
-    const intervalId = setInterval(() => {
-      const remaining = Math.max(
-        0,
-        Math.floor((data.respawnTime - Date.now()) / 1000),
-      );
-      setCountdown(remaining);
-    }, 1000);
-
-    return () => clearInterval(intervalId);
-  }, [data.respawnTime]);
-
-  // Format countdown as mm:ss
-  const formatCountdown = (seconds: number): string => {
-    const mins = Math.floor(seconds / 60);
-    const secs = seconds % 60;
-    return `${mins}:${secs.toString().padStart(2, "0")}`;
-  };
-
-  const handleRespawn = () => {
-    // Prevent multiple clicks
-    if (isRespawning) return;
-
-    // Clear timeout state on retry
-    setRespawnTimedOut(false);
-
-    // Send respawn request to server via network
-    const network = world.network as {
-      send?: (packet: string, data: unknown) => void;
-    };
-
-    if (!network) {
-      console.error("[DeathScreen] Network object is null/undefined!");
-      return;
-    }
-
-    if (!network.send) {
-      console.error("[DeathScreen] Network.send method doesn't exist!");
-      return;
-    }
-
-    // Disable button immediately to prevent spam
-    setIsRespawning(true);
-
-    try {
-      network.send("requestRespawn", {
-        playerId: world.entities?.player?.id,
-      });
-    } catch (err) {
-      console.error("[DeathScreen] Error sending packet:", err);
-      // Re-enable button on error so user can retry
-      setIsRespawning(false);
-    }
-  };
-
-  return (
-    <div
-      className="absolute inset-0 flex items-center justify-center pointer-events-auto z-[10000]"
-      style={{ backgroundColor: theme.colors.background.overlay }}
-    >
-      <div
-        className="flex flex-col items-center gap-6 max-w-md p-8 rounded-2xl backdrop-blur-md"
-        style={{
-          backgroundColor: theme.colors.background.secondary,
-          border: `2px solid ${theme.colors.state.danger}`,
-        }}
-      >
-        <div
-          className="text-4xl font-bold"
-          style={{ color: theme.colors.state.danger }}
-        >
-          Oh dear, you are dead!
-        </div>
-        <div
-          className="text-center space-y-2"
-          style={{ color: theme.colors.text.primary }}
-        >
-          <p className="text-lg">
-            Killed by:{" "}
-            <span style={{ color: theme.colors.state.danger }}>
-              {data.killedBy}
-            </span>
-          </p>
-          <p className="text-base opacity-90">
-            You have lost your items at the death location.
-          </p>
-        </div>
-        <div className="flex flex-col items-center gap-4 mt-4">
-          <button
-            onClick={handleRespawn}
-            disabled={isRespawning}
-            className="px-8 py-3 text-lg font-bold rounded-lg transition-colors border-2"
-            style={{
-              backgroundColor: isRespawning
-                ? theme.colors.text.disabled
-                : theme.colors.state.info,
-              borderColor: isRespawning
-                ? theme.colors.text.disabled
-                : theme.colors.state.info,
-              color: theme.colors.text.primary,
-              cursor: isRespawning ? "not-allowed" : "pointer",
-              opacity: isRespawning ? 0.6 : 1,
-            }}
-          >
-            {isRespawning ? "Respawning..." : "Click here to respawn"}
-          </button>
-          {respawnTimedOut && (
-            <div
-              className="text-sm text-center max-w-sm"
-              style={{ color: theme.colors.state.warning }}
-            >
-              Respawn request timed out. Please try again.
-            </div>
-          )}
-          {/* Death countdown timer */}
-          <div className="text-sm text-center max-w-sm">
-            {countdown > 0 ? (
-              <>
-                <span style={{ color: theme.colors.text.muted }}>
-                  Your items have been dropped at your death location.
-                </span>
-                <br />
-                <span
-                  className="font-bold"
-                  style={{
-                    color:
-                      countdown <= 60
-                        ? theme.colors.state.danger
-                        : theme.colors.state.warning,
-                  }}
-                >
-                  Time remaining: {formatCountdown(countdown)}
-                </span>
-              </>
-            ) : (
-              <span style={{ color: theme.colors.state.danger }}>
-                Your items have despawned!
-              </span>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
   );
 }
