+++ packages/client/src/game/hud/overlays/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Core UI Overlays
+ *
+ * Full-screen overlay components for game states like
+ * disconnection, death, and being kicked.
+ *
+ * @packageDocumentation
+ */
+
+export { Disconnected } from "./DisconnectedOverlay";
+export { KickedOverlay } from "./KickedOverlay";
+export { DeathScreen, type DeathScreenData } from "./DeathScreen";
